added functionality changing isUnusedAttendanceCode to isAvailableAttendanceCode, wrote unit tests
