+++ services/EventService.ts
@@ -24,8 +24,9 @@ export default class EventService {
   public async create(event: Event): Promise<PublicEvent> {
     const eventCreated = await this.transactions.readWrite(async (txn) => {
       const eventRepository = Repositories.event(txn);
-      const isUnusedAttendanceCode = await eventRepository.isUnusedAttendanceCode(event.attendanceCode);
-      if (!isUnusedAttendanceCode) throw new UserError('Attendance code has already been used');
+      const isAvailableAttendanceCode = await eventRepository.isAvailableAttendanceCode(event.attendanceCode,
+        event.start, event.end);
+      if (!isAvailableAttendanceCode) throw new UserError('There is a conflicting event with the same attendance code');
       if (event.start > event.end) throw new UserError('Start date after end date');
       return eventRepository.upsertEvent(EventModel.create(event));
     });
@@ -68,8 +69,11 @@ export default class EventService {
       const currentEvent = await eventRepository.findByUuid(uuid);
       if (!currentEvent) throw new NotFoundError('Event not found');
       if (changes.attendanceCode !== currentEvent.attendanceCode) {
-        const isUnusedAttendanceCode = eventRepository.isUnusedAttendanceCode(changes.attendanceCode);
-        if (!isUnusedAttendanceCode) throw new UserError('Attendance code has already been used');
+        const isAvailableAttendanceCode = eventRepository.isAvailableAttendanceCode(changes.attendanceCode,
+          changes.start, changes.end);
+        if (!isAvailableAttendanceCode) {
+          throw new UserError('There is a conflicting event with the same attendance code');
+        }
       }
       return eventRepository.upsertEvent(currentEvent, changes);
     });
